+++ CHANGELOG.md
@@ -12,6 +12,18 @@
   error with an existing stack trace, instead of creating
   a new stack trace.
 
+#### `dart:io`
+
+- **Breaking Change** [#47769](https://github.com/dart-lang/sdk/issues/47769):
+The `Platform.packageRoot` API has been removed. I had been marked deprecated in
+2018, as it doesn't work with any Dart 2.x release.
+
+#### `dart:isolate`
+
+- **Breaking Change** [#47769](https://github.com/dart-lang/sdk/issues/47769):
+The `Isolate.packageRoot` API has been removed. I had been marked deprecated in
+2018, as it doesn't work with any Dart 2.x release.
+
 ### Tools
 
 #### Dart command line