+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # 1.16.0
 
+- doc improvements for `prefer_initializing_formals`
+- updates to `secure_pubspec_urls` to check `issue_tracker` and
+  `repository` entries
 - new lint: `conditional_uri_does_not_exist`
+- performance improvements for 
+  `missing_whitespace_between_adjacent_strings`
 
 # 1.15.0